+++ lib/data/models/section.dart
@@ -100,7 +100,8 @@ class Section extends ChangeNotifier {
     }
 
     for(final original in originalItems){
-      if(!items.contains(original)){
+      //correcao de bug, no iOS ocorre um erro com as imagens, para diferenciar a origem do firebase e da internet
+      if(!items.contains(original) && (original.image as String).contains('firebase')){
         try {
           final ref = await storage.getReferenceFromUrl(
               original.image as String);
@@ -119,10 +120,13 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for(final item in items ){
-      try {
-        final ref = await storage.getReferenceFromUrl(item.image as String);
-        await ref.delete();
-      } catch (e){}
+      //correcao de bug, no iOS ocorre um erro com as imagens, para diferenciar a origem do firebase e da internet
+      if((item.image as String).contains('firebase')){
+        try {
+          final ref = await storage.getReferenceFromUrl(item.image as String);
+          await ref.delete();
+        } catch (e){}
+      }
     }
   }
